DOMA-3133 moved findAllByKey to ecmascript.utils.js
